feat: directory cleanup and reorganization on the cloudflare worker
