+++ README.md
@@ -8,6 +8,7 @@ Prettier's table formatting is cursed, hence this ignore block.
 <!-- prettier-ignore-start -->
 | | Language / Framework | Package | Latest Release |
 | :--- |:--- | :--- | :--- |
+| <img src="https://raw.githubusercontent.com/readmeio/metrics-sdks/main/docs/images/cloudflare-logo.svg" width="40" /> | Cloudflare Worker | [readmeio](https://github.com/readmeio/metrics-sdks/tree/main/packages/cloudflare-worker) | [![NPM](https://img.shields.io/npm/v/@readme/cloudflare-worker.svg)](https://npm.im/@readme/cloudflare-worker) ![Build status](https://img.shields.io/github/workflow/status/readmeio/metrics-sdks/cloudflare-worker.svg) |
 | <img src="https://raw.githubusercontent.com/readmeio/metrics-sdks/main/docs/images/node-logo.svg" width="40" /> | NodeJS (Express) | [readmeio](https://github.com/readmeio/metrics-sdks/tree/main/packages/node) | [![NPM](https://img.shields.io/npm/v/readmeio.svg)](https://npm.im/readmeio) ![Build status](https://img.shields.io/github/workflow/status/readmeio/metrics-sdks/nodejs.svg) |
 | <img src="https://raw.githubusercontent.com/readmeio/metrics-sdks/main/docs/images/php-logo.svg" width="40" /> | PHP (Laravel) | [readme/metrics](https://github.com/readmeio/metrics-sdks/tree/main/packages/php) | [![Packagist](https://img.shields.io/packagist/v/readme/metrics.svg)](https://packagist.org/packages/readme/metrics) ![Build status](https://img.shields.io/github/workflow/status/readmeio/metrics-sdks/php.svg) |
 | <img src="https://raw.githubusercontent.com/readmeio/metrics-sdks/main/docs/images/python-logo.svg" width="40" /> | Python (Django, Flask, and WSGI) | [readme-metrics](https://github.com/readmeio/metrics-sdks/tree/main/packages/python) | [![PyPI](https://img.shields.io/pypi/v/readme-metrics.svg)](https://pypi.org/project/readme-metrics/) ![Build status](https://img.shields.io/github/workflow/status/readmeio/metrics-sdks/python.svg) |
