+++ package.json
@@ -22,6 +22,7 @@
     "node": ">=14"
   },
   "workspaces": [
+    "./packages/cloudflare-worker",
     "./packages/node",
     "./packages/sdk-snippets"
   ],
