+++ packages/cloudflare-worker/.eslintignore
@@ -1,3 +1,2 @@
-**/node_modules
-coverage
-dist
+coverage/
+dist/
