+++ packages/cloudflare-worker/.eslintrc
@@ -5,6 +5,11 @@
     "browser": true,
     "worker": true
   },
+  "globals": {
+    "HOST": "writable",
+    "INSTALL_OPTIONS": "readonly",
+    "VERSION": "readonly"
+  },
   "rules": {
     "import/no-unresolved": ["error", { "ignore": ["@readme/cloudflare-worker"] }],
     "no-console": "off",
@@ -12,9 +17,18 @@
     "no-use-before-define": ["error", "nofunc"],
     "sonarjs/no-nested-template-literals": "off"
   },
-  "globals": {
-    "HOST": "writable",
-    "INSTALL_OPTIONS": "readonly",
-    "VERSION": "readonly"
+  "overrides": [
+    {
+      "files": ["src/template.js"],
+      "rules": {
+        "import/no-extraneous-dependencies": "off",
+        "node/no-extraneous-require": "off"
+      }
+    }
+  ],
+  "settings": {
+    "jest": {
+      "version": 28
+    }
   }
 }
