+++ packages/cloudflare-worker/.gitignore
@@ -1,3 +1,3 @@
-node_modules
-coverage
-.nyc_output
+coverage/
+dist/
+node_modules/
