+++ packages/cloudflare-worker/.prettierignore
@@ -1,3 +1,2 @@
-coverage
-dist
-template.js
+coverage/
+dist/
