+++ packages/cloudflare-worker/LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2018, Dom Harrington <dom@readme.io>
+Copyright (c) 2022, ReadMe
 
 Permission to use, copy, modify, and/or distribute this software for any purpose
 with or without fee is hereby granted, provided that the above copyright notice
