+++ packages/cloudflare-worker/__tests__/index.test.js
@@ -3,8 +3,8 @@ const nock = require('nock');
 const globals = require('./service-worker-globals');
 
 function requireWorker() {
-  delete require.cache[require.resolve('../')];
-  return require('..'); // eslint-disable-line global-require
+  delete require.cache[require.resolve('../src')];
+  return require('../src'); // eslint-disable-line global-require
 }
 
 describe('worker', () => {
