+++ packages/cloudflare-worker/__tests__/service-worker-globals.js
