+++ packages/cloudflare-worker/jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   moduleNameMapper: {
-    '^@readme/cloudflare-worker$': '<rootDir>/index.js',
+    '^@readme/cloudflare-worker$': '<rootDir>/src/index.js',
   },
+  testMatch: ['**/__tests__/**/(*.)+(spec|test).[jt]s?(x)', '**/?(*.)+(spec|test).[jt]s?(x)'],
 };
