+++ packages/cloudflare-worker/package.json
@@ -28,15 +28,14 @@
     "url": "git@github.com:readmeio/cloudflare-worker.git"
   },
   "devDependencies": {
-    "@readme/eslint-config": "^7.1.1",
-    "eslint": "^7.32.0",
-    "jest": "^27.2.0",
-    "nock": "^13.1.3",
-    "node-fetch": "^2.6.2",
-    "path-browserify": "^1.0.1",
-    "prettier": "^2.4.1",
-    "webpack": "^5.53.0",
-    "webpack-cli": "^4.8.0"
+    "@readme/eslint-config": "^9.0.0",
+    "eslint": "^8.21.0",
+    "jest": "^28.1.3",
+    "nock": "^13.2.9",
+    "node-fetch": "^2.6.7",
+    "prettier": "^2.7.1",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0"
   },
   "prettier": "@readme/eslint-config/prettier"
 }
